fix(pipeline_template): Fixing regression in test harness
