+++ orca-pipelinetemplate/src/test/groovy/com/netflix/spinnaker/orca/pipelinetemplate/v1schema/validator/V1TemplateConfigurationSchemaValidatorSpec.groovy
@@ -18,6 +18,7 @@ package com.netflix.spinnaker.orca.pipelinetemplate.v1schema.validator
 import com.netflix.spinnaker.orca.pipelinetemplate.v1schema.model.StageDefinition
 import com.netflix.spinnaker.orca.pipelinetemplate.v1schema.model.TemplateConfiguration
 import com.netflix.spinnaker.orca.pipelinetemplate.v1schema.model.TemplateConfiguration.PipelineDefinition
+import com.netflix.spinnaker.orca.pipelinetemplate.v1schema.model.TemplateConfiguration.TemplateSource
 import com.netflix.spinnaker.orca.pipelinetemplate.validator.Errors
 import spock.lang.Specification
 import spock.lang.Subject
@@ -78,7 +79,10 @@ class V1TemplateConfigurationSchemaValidatorSpec extends Specification {
     def errors = new Errors()
     def templateConfiguration = new TemplateConfiguration(
       schema: "1",
-      pipeline: new PipelineDefinition(application: 'myapp'),
+      pipeline: new PipelineDefinition(
+        application: 'myapp',
+        template: new TemplateSource()
+      ),
       stages: [
         new StageDefinition(
           id: 'foo',
@@ -93,6 +97,7 @@ class V1TemplateConfigurationSchemaValidatorSpec extends Specification {
 
     then:
     if (hasErrors) {
+      errors.hasErrors(true)
       errors.errors[0].message == "A configuration-defined stage should have either dependsOn or an inject rule defined"
       errors.errors[0].location == 'configuration:stages.foo'
     } else {
